MTL-2406 Add `binutils` packages
Adds `binutils-devel` and `binutils-gold` to resolve arm64 build issues for
some languages that require `ld`.

These packages come from the `sle-module-development-tools` and
`sle-module-desktop-applications` modules.

The SLE_BCI repository has the matching version of `binutils` dependency
but since it's a lower priority than the other repos zypper does not opt
to install it. The easy way around this is to modify `SLE_BCI` to have
the same priority as the other repos.